+++ packages/utils/src/pickProProps/index.tsx
@@ -1,6 +1,6 @@
-const proFieldProps = `valueType request mode plain renderFormItem render text formItemProps valueEnum`;
+const proFieldProps = `valueType request plain renderFormItem render text formItemProps valueEnum`;
 
-const proFormProps = `fieldProps secondary colSize groupProps contentRender submitterProps submitter`;
+const proFormProps = `fieldProps isDefaultDom secondary colSize groupProps contentRender submitterProps submitter`;
 
 export default function pickProProps(props: object) {
   const propList = `${proFieldProps} ${proFormProps}`.split(/[\s\n]+/);