+++ packages/layout/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.4.9](https://github.com/ant-design/pro-components/compare/@ant-design/pro-layout@6.4.8...@ant-design/pro-layout@6.4.9) (2020-08-19)
+
+**Note:** Version bump only for package @ant-design/pro-layout
+
 ## [6.4.8](https://github.com/ant-design/pro-components/compare/@ant-design/pro-layout@6.4.7...@ant-design/pro-layout@6.4.8) (2020-08-17)
 
 **Note:** Version bump only for package @ant-design/pro-layout